rewrite tests for CommandCaller to allow access to private variables
