+++ test/src/threads/command_caller.h
@@ -14,20 +14,20 @@ const std::string error_str   = "error";
 class CommandCallerMock : public CommandCaller
 {
 public:
-  FRIEND_TEST(CommandCallerTest, TEST_UNCONF);
-  FRIEND_TEST(CommandCallerTest, TEST_LOG);
-  FRIEND_TEST(CommandCallerTest, TEST_STATUS);
-  FRIEND_TEST(CommandCallerTest, TEST_CHANGE_STATUS_EE);
-  FRIEND_TEST(CommandCallerTest, TEST_CHANGE_STATUS_CC);
-  FRIEND_TEST(CommandCallerTest, TEST_CHANGE_STATUS_CE);
-  FRIEND_TEST(CommandCallerTest, TEST_CHANGE_STATUS_EC_SUCCESS);
-  FRIEND_TEST(CommandCallerTest, TEST_CHANGE_STATUS_CE_FAIL);
-
   MOCK_METHOD(std::string, call_command, (const std::vector<std::string> &, const std::string &), (override));
   MOCK_METHOD(CommandCaller::results, call_command, (const std::vector<std::string> &), (override));
 
   CommandCallerMock::results result_success = { 0, success_str, error_str };
   CommandCallerMock::results result_error   = { 1, success_str, error_str };
+};
+
+class CommandCallerTest : public ::testing::Test
+{
+protected:
+  CommandCallerTest() = default;
+  virtual void SetUp() {}
+
+  CommandCallerMock tester;
 };
 
 #endif
